refine getEZForPath rpc code

+++ depends/libhdfs3/src/client/hdfs.h
@@ -487,7 +487,7 @@ int hdfsSetReplication(hdfsFS fs, const char * path, int16_t replication);
 typedef struct {
     int mSuite; /* the suite of encryption zone */
     int mCryptoProtocolVersion; /* the version of crypto protocol */
-    int mId; /* the id of encryption zone */
+    int64_t mId; /* the id of encryption zone */
     char * mPath; /* the path of encryption zone */
     char * mKeyName; /* the key name of encryption zone */
 } hdfsEncryptionZoneInfo;
